inspect Xcode info to determine run destination
each new version of Xcode contains a new batch of simulators. using hard-coded run destinations is brittle, because these change over time.

instead, query the local Xcode installation for information about the simulators. Xcode doesn't return an easily parseable format, so we're making some assumptions about the order of the simulators it returns (it seems like the last item in the list for a given platform is always for the latest SDK)

this will hopefully make Travis less brittle and will make it possible for us as a team to all be on different Xcode installations and still use the same releaseMe.sh script